refactor(otelcol): decouple debug_metrics block from otelcol package
