+++ internal/component/otelcol/config/config_debug_metrics.go
