+++ internal/component/otelcol/exporter/exporter_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/grafana/alloy/internal/component"
 	"github.com/grafana/alloy/internal/component/otelcol"
+	otelcolCfg "github.com/grafana/alloy/internal/component/otelcol/config"
 	"github.com/grafana/alloy/internal/component/otelcol/exporter"
 	"github.com/grafana/alloy/internal/runtime/componenttest"
 	"github.com/grafana/alloy/internal/util"
@@ -137,8 +138,8 @@ func (fa fakeExporterArgs) Exporters() map[otelcomponent.DataType]map[otelcompon
 	return nil
 }
 
-func (fe fakeExporterArgs) DebugMetricsConfig() otelcol.DebugMetricsArguments {
-	var dma otelcol.DebugMetricsArguments
+func (fe fakeExporterArgs) DebugMetricsConfig() otelcolCfg.DebugMetricsArguments {
+	var dma otelcolCfg.DebugMetricsArguments
 	dma.SetToDefault()
 	return dma
 }
