+++ internal/component/otelcol/exporter/otlphttp/otlphttp_test.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/grafana/alloy/internal/component/otelcol"
+	otelcolCfg "github.com/grafana/alloy/internal/component/otelcol/config"
 	"github.com/grafana/alloy/internal/component/otelcol/exporter/otlphttp"
 	"github.com/grafana/alloy/internal/runtime/componenttest"
 	"github.com/grafana/alloy/internal/runtime/logging/level"
@@ -119,7 +120,7 @@ func TestDebugMetricsConfig(t *testing.T) {
 	tests := []struct {
 		testName string
 		alloyCfg string
-		expected otelcol.DebugMetricsArguments
+		expected otelcolCfg.DebugMetricsArguments
 	}{
 		{
 			testName: "default",
@@ -128,7 +129,7 @@ func TestDebugMetricsConfig(t *testing.T) {
 				endpoint = "http://tempo:4317"
 			}
 			`,
-			expected: otelcol.DebugMetricsArguments{
+			expected: otelcolCfg.DebugMetricsArguments{
 				DisableHighCardinalityMetrics: true,
 			},
 		},
@@ -142,7 +143,7 @@ func TestDebugMetricsConfig(t *testing.T) {
 				disable_high_cardinality_metrics = false
 			}
 			`,
-			expected: otelcol.DebugMetricsArguments{
+			expected: otelcolCfg.DebugMetricsArguments{
 				DisableHighCardinalityMetrics: false,
 			},
 		},
@@ -156,7 +157,7 @@ func TestDebugMetricsConfig(t *testing.T) {
 				disable_high_cardinality_metrics = true
 			}
 			`,
-			expected: otelcol.DebugMetricsArguments{
+			expected: otelcolCfg.DebugMetricsArguments{
 				DisableHighCardinalityMetrics: true,
 			},
 		},
