+++ internal/component/otelcol/processor/filter/filter.go
@@ -3,6 +3,7 @@ package filter
 import (
 	"github.com/grafana/alloy/internal/component"
 	"github.com/grafana/alloy/internal/component/otelcol"
+	otelcolCfg "github.com/grafana/alloy/internal/component/otelcol/config"
 	"github.com/grafana/alloy/internal/component/otelcol/processor"
 	"github.com/grafana/alloy/internal/featuregate"
 	"github.com/mitchellh/mapstructure"
@@ -37,7 +38,7 @@ type Arguments struct {
 	Output *otelcol.ConsumerArguments `alloy:"output,block"`
 
 	// DebugMetrics configures component internal metrics. Optional.
-	DebugMetrics otelcol.DebugMetricsArguments `alloy:"debug_metrics,block,optional"`
+	DebugMetrics otelcolCfg.DebugMetricsArguments `alloy:"debug_metrics,block,optional"`
 }
 
 var (
@@ -114,6 +115,6 @@ func (args Arguments) NextConsumers() *otelcol.ConsumerArguments {
 }
 
 // DebugMetricsConfig implements processor.Arguments.
-func (args Arguments) DebugMetricsConfig() otelcol.DebugMetricsArguments {
+func (args Arguments) DebugMetricsConfig() otelcolCfg.DebugMetricsArguments {
 	return args.DebugMetrics
 }
