+++ internal/component/otelcol/processor/k8sattributes/k8sattributes.go
@@ -4,6 +4,7 @@ package k8sattributes
 import (
 	"github.com/grafana/alloy/internal/component"
 	"github.com/grafana/alloy/internal/component/otelcol"
+	otelcolCfg "github.com/grafana/alloy/internal/component/otelcol/config"
 	"github.com/grafana/alloy/internal/component/otelcol/processor"
 	"github.com/grafana/alloy/internal/featuregate"
 	"github.com/mitchellh/mapstructure"
@@ -43,7 +44,7 @@ type Arguments struct {
 	Output *otelcol.ConsumerArguments `alloy:"output,block"`
 
 	// DebugMetrics configures component internal metrics. Optional.
-	DebugMetrics otelcol.DebugMetricsArguments `alloy:"debug_metrics,block,optional"`
+	DebugMetrics otelcolCfg.DebugMetricsArguments `alloy:"debug_metrics,block,optional"`
 }
 
 // SetToDefault implements syntax.Defaulter.
@@ -123,6 +124,6 @@ func (args Arguments) NextConsumers() *otelcol.ConsumerArguments {
 }
 
 // DebugMetricsConfig implements processor.Arguments.
-func (args Arguments) DebugMetricsConfig() otelcol.DebugMetricsArguments {
+func (args Arguments) DebugMetricsConfig() otelcolCfg.DebugMetricsArguments {
 	return args.DebugMetrics
 }
