+++ internal/component/otelcol/processor/processor_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/grafana/alloy/internal/component"
 	"github.com/grafana/alloy/internal/component/otelcol"
+	otelcolCfg "github.com/grafana/alloy/internal/component/otelcol/config"
 	"github.com/grafana/alloy/internal/component/otelcol/internal/fakeconsumer"
 	"github.com/grafana/alloy/internal/component/otelcol/processor"
 	"github.com/grafana/alloy/internal/runtime/componenttest"
@@ -171,8 +172,8 @@ func (fa fakeProcessorArgs) NextConsumers() *otelcol.ConsumerArguments {
 	return fa.Output
 }
 
-func (fa fakeProcessorArgs) DebugMetricsConfig() otelcol.DebugMetricsArguments {
-	var dma otelcol.DebugMetricsArguments
+func (fa fakeProcessorArgs) DebugMetricsConfig() otelcolCfg.DebugMetricsArguments {
+	var dma otelcolCfg.DebugMetricsArguments
 	dma.SetToDefault()
 	return dma
 }
